+++ tests/test_api.py
@@ -1,6 +1,5 @@
 import json
-import os
-from datetime import datetime
+from datetime import datetime, timedelta
 from pathlib import Path
 
 import pytest
@@ -71,9 +70,8 @@ async def test_api_discharge_docs(monkeypatch):
         test_data = [PatientFile(**item) for item in test_data]
 
     monkeypatch.setattr(app, "client", MockAzureOpenAI())
-    output = await process_and_generate_discharge_docs(
-        test_data, FakeDB(), os.getenv("X_API_KEY")
-    )
+    monkeypatch.setenv("X_API_KEY_generate", "test")
+    output = await process_and_generate_discharge_docs(test_data, FakeDB(), "test")
     assert output["message"] == "Success"
     assert isinstance(output["discharge_letter"], str)
 
@@ -82,9 +80,8 @@ async def test_api_discharge_docs(monkeypatch):
 async def test_api_remove_discharge_docs(monkeypatch):
     """Test the remove_old_discharge_docs endpoint in the API."""
     monkeypatch.setattr(app, "client", MockAzureOpenAI())
-    output = await remove_old_discharge_docs(
-        datetime(2023, 1, 1), FakeDB(), os.getenv("X_API_KEY")
-    )
+    monkeypatch.setenv("X_API_KEY_remove", "test")
+    output = await remove_old_discharge_docs(datetime(2023, 1, 1), FakeDB(), "test")
     assert output == {"message": "Success"}
 
 
@@ -100,6 +97,7 @@ async def test_api_wrong_api_key(monkeypatch):
         pytest.fail(f"JSON data does not match PatientFile schema: {e}")
 
     monkeypatch.setattr(app, "client", MockAzureOpenAI())
+    monkeypatch.setenv("X_API_KEY_generate", "test")
 
     try:
         await process_and_generate_discharge_docs(test_data, FakeDB(), "wrong_api_key")
@@ -109,10 +107,362 @@ async def test_api_wrong_api_key(monkeypatch):
 
 
 @pytest.mark.asyncio
-async def test_api_get_discharge_docs(monkeypatch):
-    """Test the process_and_generate_discharge_docs endpoint in the API."""
+async def test_api_retrieve_discharge_docs(monkeypatch):
+    """Test the retrieve endpoint in the API."""
     monkeypatch.setattr(app, "client", MockAzureOpenAI())
-    output = await app.retrieve_discharge_doc(
-        "test_patient_id", FakeDB(), os.getenv("X_API_KEY")
-    )
+    monkeypatch.setenv("X_API_KEY_retrieve", "test")
+
+    output = await app.retrieve_discharge_doc("test_patient_id", FakeDB(), "test")
     assert isinstance(output, str)
+
+
+@pytest.mark.asyncio
+async def test_api_retrieve_discharge_doc_success(monkeypatch):
+    """Test retrieving a successful discharge letter for a patient."""
+    mock_data = [
+        (
+            "Discharge letter content",
+            "enc123",
+            1,
+            "Success",
+            datetime.now() - timedelta(days=8),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            2,
+            "Success",
+            datetime.now() - timedelta(days=7),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            3,
+            "Success",
+            datetime.now() - timedelta(days=6),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            4,
+            "Success",
+            datetime.now() - timedelta(days=5),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            5,
+            "Success",
+            datetime.now() - timedelta(days=4),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            6,
+            "Success",
+            datetime.now() - timedelta(days=3),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            7,
+            "Success",
+            datetime.now() - timedelta(days=2),
+        ),
+        (
+            "Discharge Letter content",
+            "enc123",
+            8,
+            "Success",
+            datetime.now() - timedelta(days=1),
+        ),
+        (
+            "Most Recent Successful Discharge Letter",
+            "enc123",
+            9,
+            "Success",
+            datetime.now(),
+        ),
+    ]
+    mock_data = sorted(mock_data, key=lambda x: x[2], reverse=True)
+
+    class FakeExecuteSuccess(FakeExecute):
+        def fetchall(self):
+            print("requested fetchall with results...")
+            return mock_data
+
+    class FakeDBWithResults(FakeDB):
+        def execute(self, stmt):
+            print(f"{stmt} executed with fetchall data...")
+            return FakeExecuteSuccess()
+
+    monkeypatch.setattr(app, "client", MockAzureOpenAI())
+    monkeypatch.setenv("X_API_KEY_retrieve", "test")
+    output = await app.retrieve_discharge_doc(
+        "test_patient_id", FakeDBWithResults(), "test"
+    )
+
+    assert isinstance(output, str)
+    assert "Most Recent Successful Discharge Letter" in output
+
+
+@pytest.mark.asyncio
+async def test_api_retrieve_discharge_doc_no_results(monkeypatch):
+    """Test retrieving discharge docs with no results for a patient."""
+    monkeypatch.setattr(app, "client", MockAzureOpenAI())
+    monkeypatch.setenv("X_API_KEY_retrieve", "test")
+
+    output = await app.retrieve_discharge_doc("test_patient_id", FakeDB(), "test")
+
+    assert isinstance(output, str)
+    assert "Er is geen ontslagbrief in de database gevonden" in output
+
+
+@pytest.mark.asyncio
+async def test_api_retrieve_discharge_doc_older_letter(monkeypatch):
+    """Test retrieving discharge docs with a LengthError."""
+    mock_data = [
+        (
+            "Discharge letter content",
+            "enc123",
+            1,
+            "Success",
+            datetime.now() - timedelta(days=8),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            2,
+            "Success",
+            datetime.now() - timedelta(days=7),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            3,
+            "Success",
+            datetime.now() - timedelta(days=6),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            4,
+            "Success",
+            datetime.now() - timedelta(days=5),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            5,
+            "Success",
+            datetime.now() - timedelta(days=4),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            6,
+            "Success",
+            datetime.now() - timedelta(days=3),
+        ),
+        (
+            "Discharge letter content",
+            "enc123",
+            7,
+            "Success",
+            datetime.now() - timedelta(days=2),
+        ),
+        (
+            "Most Recent Successful Discharge Letter",
+            "enc123",
+            8,
+            "Success",
+            datetime.now() - timedelta(days=1),
+        ),
+        (
+            "Too Long Discharge Letter",
+            "enc123",
+            9,
+            "LengthError",
+            datetime.now(),
+        ),
+    ]
+    mock_data = sorted(mock_data, key=lambda x: x[2], reverse=True)
+
+    class FakeExecuteLengthError(FakeExecute):
+        def fetchall(self):
+            print("requested fetchall with LengthError...")
+            return mock_data
+
+    class FakeDBWithLengthError(FakeDB):
+        def execute(self, stmt):
+            print(f"{stmt} executed with LengthError data...")
+            return FakeExecuteLengthError()
+
+    monkeypatch.setattr(app, "client", MockAzureOpenAI())
+    monkeypatch.setenv("X_API_KEY_retrieve", "test")
+
+    output = await app.retrieve_discharge_doc(
+        "test_patient_id", FakeDBWithLengthError(), "test"
+    )
+
+    assert isinstance(output, str)
+    assert "Most Recent Successful Discharge Letter" in output
+
+
+@pytest.mark.asyncio
+async def test_api_retrieve_discharge_doc_outdated(monkeypatch):
+    """Test retrieving an outdated discharge letter."""
+    mock_data = [
+        (
+            "Discharge Letter Content",
+            "enc123",
+            1,
+            "Success",
+            datetime.now() - timedelta(days=8),
+        ),
+        (
+            "Discharge Letter Content",
+            "enc123",
+            2,
+            "Success",
+            datetime.now() - timedelta(days=7),
+        ),
+        (
+            "Discharge Letter Content",
+            "enc123",
+            3,
+            "Success",
+            datetime.now() - timedelta(days=6),
+        ),
+        (
+            "Discharge Letter Content",
+            "enc123",
+            4,
+            "Success",
+            datetime.now() - timedelta(days=5),
+        ),
+        (
+            "Discharge Letter Content",
+            "enc123",
+            5,
+            "Success",
+            datetime.now() - timedelta(days=4),
+        ),
+        (
+            "Discharge Letter Content",
+            "enc123",
+            6,
+            "Success",
+            datetime.now() - timedelta(days=3),
+        ),
+        (
+            "Most Recent Discharge Letter Content",
+            "enc123",
+            7,
+            "Success",
+            datetime.now() - timedelta(days=2),
+        ),
+        (
+            "",
+            "enc123",
+            8,
+            "GeneralError",
+            datetime.now() - timedelta(days=1),
+        ),
+        (
+            "",
+            "enc123",
+            9,
+            "GeneralError",
+            datetime.now(),
+        ),
+    ]
+    mock_data = sorted(mock_data, key=lambda x: x[2], reverse=True)
+
+    class FakeExecuteOutdated(FakeExecute):
+        def fetchall(self):
+            print("requested fetchall with outdated data...")
+            return mock_data
+
+    class FakeDBWithOutdatedData(FakeDB):
+        def execute(self, stmt):
+            print(f"{stmt} executed with outdated data...")
+            return FakeExecuteOutdated()
+
+    monkeypatch.setattr(app, "client", MockAzureOpenAI())
+    monkeypatch.setenv("X_API_KEY_retrieve", "test")
+
+    output = await app.retrieve_discharge_doc(
+        "test_patient_id", FakeDBWithOutdatedData(), "test"
+    )
+
+    assert isinstance(output, str)
+    assert "NB Let erop dat deze brief niet afgelopen nacht is gegenereerd" in output
+    assert "Most Recent Discharge Letter Content" in output
+
+
+@pytest.mark.asyncio
+async def test_api_retrieve_discharge_doc_wrong_api_key(monkeypatch):
+    """Test retrieving discharge docs with an incorrect API key."""
+    monkeypatch.setattr(app, "client", MockAzureOpenAI())
+    monkeypatch.setenv("X_API_KEY_retrieve", "test")
+
+    try:
+        await app.retrieve_discharge_doc("test_patient_id", FakeDB(), "wrong_api_key")
+    except HTTPException as e:
+        assert e.status_code == 403
+        assert e.detail == "You are not authorized to access this endpoint"
+
+
+@pytest.mark.asyncio
+async def test_api_retrieve_discharge_doc__general_error_then_success(
+    monkeypatch,
+):
+    """Test retrieving discharge docs with first a GeneralError, then a success"""
+
+    mock_data = [
+        ("", "enc123", 1, "GeneralError", datetime.now() - timedelta(days=8)),
+        ("", "enc123", 2, "GeneralError", datetime.now() - timedelta(days=7)),
+        ("", "enc123", 3, "GeneralError", datetime.now() - timedelta(days=6)),
+        ("", "enc123", 4, "GeneralError", datetime.now() - timedelta(days=5)),
+        ("", "enc123", 5, "GeneralError", datetime.now() - timedelta(days=4)),
+        ("", "enc123", 6, "GeneralError", datetime.now() - timedelta(days=3)),
+        ("", "enc123", 7, "GeneralError", datetime.now() - timedelta(days=2)),
+        (
+            "Successful Discharge Letter",
+            "enc123",
+            8,
+            "Success",
+            datetime.now() - timedelta(days=1),
+        ),
+        (
+            "Most Recent Successful Discharge Letter",
+            "enc123",
+            9,
+            "Success",
+            datetime.now(),
+        ),
+    ]
+    mock_data = sorted(mock_data, key=lambda x: x[2], reverse=True)
+
+    class FakeExecuteEveryDayEntry(FakeExecute):
+        def fetchall(self):
+            print(
+                "requested fetchall with entries for every day over more than a week..."
+            )
+            return mock_data
+
+    class FakeDBEveryDayEntry(FakeDB):
+        def execute(self, stmt):
+            print(f"{stmt} executed with entries for every day...")
+            return FakeExecuteEveryDayEntry()
+
+    # Use the mocked DB session and mock Azure client
+    monkeypatch.setattr(app, "client", MockAzureOpenAI())
+    monkeypatch.setenv("X_API_KEY_retrieve", "test")
+    output = await app.retrieve_discharge_doc(
+        "test_patient_id", FakeDBEveryDayEntry(), "test"
+    )
+
+    assert isinstance(output, str)
+    assert "Most Recent Successful Discharge Letter" in output